Fix chatroomactivity

+++ app/src/main/java/com/example/tinder/Model/ChatRow.java
@@ -1,19 +1,19 @@
 package com.example.tinder.Model;
 
 public class ChatRow {
-    public float id;
+    public long id;
     public String name;
     public String lastMessage;
     public String picture;
 
-    public ChatRow(float id, String name, String lastMessage, String picture) {
+    public ChatRow(long id, String name, String lastMessage, String picture) {
         this.id = id;
         this.name = name;
         this.lastMessage = lastMessage;
         this.picture = picture;
     }
 
-    public float getId() {
+    public long getId() {
         return id;
     }
